indentation

+++ client/src/components/admin/AdminUserDelete.vue
@@ -25,10 +25,10 @@ export default {
   }),
   methods: {
     async deleteUser() {
-      // if (this.confirmDelete == "delete") {
-      const response = await deleteUser(this.id);
-      return response;
-      // }
+      if (this.confirmDelete == "delete") {
+        const response = await deleteUser(this.id);
+        return response;
+      }
     },
   },
   props: {
